Split fabric8-karaf package to provide single functionalities
