+++ components/fabric8-karaf/fabric8-karaf-blueprint/src/main/java/io/fabric8/karaf/blueprint/KubernetesPropertyEvaluator.java
@@ -13,7 +13,7 @@
  *  implied.  See the License for the specific language governing
  *  permissions and limitations under the License.
  */
-package io.fabric8.karaf;
+package io.fabric8.karaf.blueprint;
 
 import java.util.Dictionary;
 import java.util.Map;
