+++ components/fabric8-karaf/fabric8-karaf-blueprint/src/main/java/io/fabric8/karaf/blueprint/Support.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package io.fabric8.karaf;
+package io.fabric8.karaf.blueprint;
 
 public class Support {
