+++ components/fabric8-karaf/fabric8-karaf-blueprint/src/main/java/io/fabric8/karaf/blueprint/SysPropertiesFunction.java
@@ -14,10 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.karaf;
-
-import static io.fabric8.karaf.Support.after;
-import static io.fabric8.karaf.Support.before;
+package io.fabric8.karaf.blueprint;
 
 /**
  * A {@link PropertiesFunction} that lookup the property value from
@@ -36,8 +33,8 @@ public class SysPropertiesFunction implements PropertiesFunction {
         String defaultValue = null;
 
         if (remainder.contains(":")) {
-            key = before(remainder, ":");
-            defaultValue = after(remainder, ":");
+            key = Support.before(remainder, ":");
+            defaultValue = Support.after(remainder, ":");
         }
 
         String value = System.getProperty(key);
