+++ components/fabric8-karaf/fabric8-karaf-features/src/main/resources/feature.xml
@@ -1,0 +1,27 @@
+<!--
+  ~ Copyright 2005-2016 Red Hat, Inc.
+  ~
+  ~ Red Hat licenses this file to you under the Apache License, version
+  ~ 2.0 (the "License"); you may not use this file except in compliance
+  ~ with the License.  You may obtain a copy of the License at
+  ~
+  ~    http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or
+  ~ implied.  See the License for the specific language governing
+  ~ permissions and limitations under the License.
+  -->
+<features xmlns="http://karaf.apache.org/xmlns/features/v1.2.0" name="fabric8-karaf-features-${project.version}">
+
+  <feature name="fabric8-karaf-blueprint" description="Fabric8 Karaf Blueprint" version="${project.version}">
+    <bundle dependency="true">mvn:commons-codec/commons-codec/${commons-codec.version}</bundle>
+    <bundle start-level="60">mvn:io.fabric8/fabric8-karaf-blueprint/${project.version}</bundle>
+  </feature>
+
+  <feature name="fabric8-karaf-cm" description="Fabric8 Karaf ConfigAdmin" version="${project.version}">
+    <bundle start-level="60">mvn:io.fabric8/fabric8-karaf-cm/${project.version}</bundle>
+  </feature>
+
+</features>
